feat(lib): session readers 升格 lib API（#228 F3——主 repo bro hook 消費面）
