+++ .psc_tmp/moc-85g6Zv/archive/fragments/2026-06/claude__sess-e2e__000.md
@@ -1,0 +1,17 @@
+---
+memory_layer: inbox
+project: paulshaclaw
+source_agent: claude
+source_session: sess-e2e
+source_artifact: research
+captured_at: 2026-05-31T00:00:00Z
+session_title: ""
+provenance:
+  repo: paulshaclaw
+  commit: c
+  path: docs/x.md
+fragment_index: 0
+parent_session_ref: claude:sess-e2e
+---
+# Topic A
+alpha body
