+++ .psc_tmp/moc-85g6Zv/archive/sessions/2026-05/claude__sess-e2e.md
@@ -1,0 +1,16 @@
+---
+memory_layer: inbox
+project: paulshaclaw
+source_agent: claude
+source_session: sess-e2e
+source_artifact: research
+captured_at: "2026-05-31T00:00:00Z"
+provenance:
+  repo: paulshaclaw
+  commit: c
+  path: docs/x.md
+---
+# Topic A
+alpha body
+# Topic B
+beta body
