+++ .psc_tmp/moc-85g6Zv/knowledge/common-sense-moc.md
@@ -1,0 +1,8 @@
+---
+memory_layer: moc
+moc_kind: common-sense
+generated_ts: 2026-06-03T05:00:00Z
+---
+# Common-sense MOC
+
+
