+++ .psc_tmp/moc-85g6Zv/knowledge/paulshaclaw-moc.md
@@ -1,0 +1,11 @@
+---
+memory_layer: moc
+moc_kind: project
+generated_ts: 2026-06-03T05:00:00Z
+project: paulshaclaw
+---
+# paulshaclaw MOC
+
+- None
+  - [[topic-a--sl-f53f3434e9e62f53|None]] — research
+  - [[topic-b--sl-fab0c9f710b4ecef|None]] — research
