+++ .psc_tmp/moc-85g6Zv/knowledge/paulshaclaw/topic-a--sl-f53f3434e9e62f53.md
@@ -1,0 +1,29 @@
+---
+phase: research
+project: paulshaclaw
+slice_id: sl-f53f3434e9e62f53
+artifact_kind: research
+version: 1
+created_at: "2026-05-31 00:00:00+00:00"
+created_by: claude
+source_session: sess-e2e
+gate_required: false
+checksum: ed3d31ad1856a8a2b2d1723e62181e958eefd4eb40f8a37521ec154e2ca95a05
+memory_layer: knowledge
+source_agent: claude
+captured_at: "2026-05-31 00:00:00+00:00"
+supersedes: []
+distilled_from: "claude:sess-e2e"
+fragment_ref: claude__sess-e2e__000
+session_title: 
+provenance:
+  repo: paulshaclaw
+  commit: c
+  path: docs/x.md
+title: topic-a
+aliases:
+  - topic-a
+related: []
+---
+# Topic A
+alpha body
