+++ .psc_tmp/moc-85g6Zv/knowledge/paulshaclaw/topic-b--sl-fab0c9f710b4ecef.md
@@ -1,0 +1,29 @@
+---
+phase: research
+project: paulshaclaw
+slice_id: sl-fab0c9f710b4ecef
+artifact_kind: research
+version: 1
+created_at: "2026-05-31 00:00:00+00:00"
+created_by: claude
+source_session: sess-e2e
+gate_required: false
+checksum: 7317cb2f6185d269da1bb268bed69af5e9e8f8efc5df97f5b8ca44f924866644
+memory_layer: knowledge
+source_agent: claude
+captured_at: "2026-05-31 00:00:00+00:00"
+supersedes: []
+distilled_from: "claude:sess-e2e"
+fragment_ref: claude__sess-e2e__001
+session_title: 
+provenance:
+  repo: paulshaclaw
+  commit: c
+  path: docs/x.md
+title: topic-b
+aliases:
+  - topic-b
+related: []
+---
+# Topic B
+beta body
