+++ .psc_tmp/moc-85g6Zv/runtime/ledger/dream.jsonl
@@ -1,0 +1,1 @@
+{"dream_config_hash":"f831540d:780b3dd7","dry_run":false,"errors":[],"passes":{"atomize":{"config_hash":"f831540dddfc375708018a7ca66b279b60ef200a48647b7d07d949d2a8a3a676","dry_run":false,"noise_dropped":0,"skipped":0,"slices":2,"split_sessions":1},"janitor":{"config_hash":"780b3dd7315a8ed1863493f3874a62012345d4f8291394987b78d09767cc1b18","decayed":0,"dry_run":false,"lint":{"raw_remote_key":0,"untitled":0},"reactivated":0,"scanned":2,"skipped":0,"unchanged":2},"moc":{"faceout":true,"index_stats":{"paulshaclaw":{"exclude_rate":0.0,"excluded":0,"indexed":2}},"indexed":true,"linked":2,"mocs":true,"renamed":true}},"run_id":"dream-2026-06-03T05:00:00Z","status":"ok","ts":"2026-06-03T05:00:00Z"}
