+++ changelog.d/lib-session-readers.md
@@ -1,0 +1,2 @@
+### Added
+- `paulsha_hippo.lib.session_readers`：`read_codex_rollout`/`read_copilot_history` 升格 lib API（hippo importer + paulshaclaw bro hook 兩使用者；adapters.base 保留 re-export）
